ml_project/requirements.txt
pandas-profiling==3.4.0
hydra-core==1.2.0
pandas==1.5.1
numpy==1.23.4
scikit-learn==1.1.3
dill==0.3.6